refactor: split profile settings pages
